ketabrooz-bot/requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
apscheduler==3.10.4
pillow==10.1.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
requests==2.31.0
aiohttp==3.9.1
pymupdf==1.23.8

